requirements.txt
Django==4.1.1
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
psycopg2-binary==2.9.3
